requirements.txt
# requirements.txt
# Pinned dependencies for INTELLISEARCH (stable, minimal set).
# Packages are organized by priority for installation order.

# ============================================================================
# CORE PACKAGES (install first)
# ============================================================================
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.8.0
nest_asyncio>=1.5.6

# ============================================================================
# LANGCHAIN CORE PACKAGES (critical for functionality)
# ============================================================================
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0

# LangGraph workflow engine
langgraph>=0.1.0

# ============================================================================
# LLM PROVIDER PACKAGES (provider-specific)
# ============================================================================
# Google AI
google-genai>=1.0.0
langchain-google-genai>=1.0.0,<2.0.0

# Anthropic
anthropic>=0.25.0
langchain-anthropic>=0.1.0

# Together AI
together>=1.0.0
langchain-together>=0.1.0

# Voyage AI (optional)
# voyageai>=0.2.0
# langchain-voyageai>=0.1.0

# ============================================================================
# WEB SCRAPING & HTTP
# ============================================================================
aiohttp>=3.9.0
beautifulsoup4>=4.12.2
requests-html>=0.10.0
lxml>=5.0.0
trafilatura>=1.7.2

# ============================================================================
# DOCUMENT PROCESSING
# ============================================================================
pymupdf>=1.23.0
pypdf>=4.0.0
fpdf2>=2.8.0

# ============================================================================
# SEARCH & RANKING
# ============================================================================
rank_bm25>=0.2.2

# ============================================================================
# UTILITIES & PERFORMANCE
# ============================================================================
ratelimit>=2.2.1
rich>=13.3.4

# ============================================================================
# VECTOR DATABASES (optional but recommended)
# ============================================================================
# chromadb>=0.4.0
faiss-cpu>=1.7.4  # For vector similarity search (install with pip or conda)

# ============================================================================
# DEVELOPMENT & TESTING (optional)
# ============================================================================
# pytest>=7.4.0

# Notes:
# - Install packages in the order listed for best compatibility
# - Some packages (faiss-cpu, chromadb) may require conda or special handling
# - Version ranges prevent conflicts between LangChain packages
# - Comment out optional packages if not needed to reduce installation time